==> 3_analyses_FFD_LFD_MeanFD.Rmd <==
---
title: Genetic differentiation on flowering time in Cerastium fontanum using a reciprocal transplant experiment
author: "Alicia Valdés"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  pdf_document:
    latex_engine: xelatex
    toc: yes
    toc_depth: 4
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
subtitle: Analyses with FFD
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
options(tibble.width = Inf)
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load packages, include=FALSE}
library(tidyverse)
library(readxl)
library(RColorBrewer)
library(knitr)
library(ggthemes)
library(glmmTMB)
library(lme4)
library(lmerTest)
library(car)
library(sjPlot)
library(ggeffects)
library(performance)
library(partR2)
```

```{r Define ggplot themes and palettes, include=FALSE}
my_theme <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(legend.position="none")+theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
my_theme_legend <- function(){
  theme_base()+theme(plot.background=element_rect(fill="white", colour=NA))+
  theme(text=element_text(family="serif"))+
  theme(plot.title = element_text(hjust =-0.06))
}
myPalette <- colorRampPalette(brewer.pal(11, "YlOrRd"))
```

# Read clean data from .csv files

```{r}
data_transplants <- read_csv("data/clean/data_transplants.csv")
```

# Distributions

```{r}
ggplot(data_transplants,aes(x=FFD_corr))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=LFD_corr))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=MeanFD))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=log(FFD_corr)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=log(LFD_corr)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=log(MeanFD)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=sqrt(FFD_corr)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=sqrt(LFD_corr)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
ggplot(data_transplants,aes(x=sqrt(MeanFD)))+
  geom_histogram(color="black",fill="white")+
  my_theme()
```

None of the distribution is looking super normal, but what else can we use?

# TO DO: Check model diagnostics

# Models FFD (OLD but USE?)

## Temp mother + Temp father

```{r}
data_transplants$heat_zone <- factor(data_transplants$heat_zone, 
                                     levels=c("Cold", "Intermediate", "Hot"))
```

```{r}
model_FFD1<-lmer(FFD_corr~(temp_mother+temp_father)*heat_zone+
                   (1|mother)+(1|father)+(1|crossing)+(1|plot),data_transplants)
model_FFD2<-lmer(FFD_corr~(temp_mother+temp_father)*temp+
                   (1|mother)+(1|father)+(1|crossing)+(1|plot),data_transplants)
# Both models with singular fit
# Because of low variance in some of the random effectS?
```

```{r}
tab_model(model_FFD1,model_FFD2,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("heat_zone","temp"),
          title="Models FFD")
```

Save models as HTML table

```{r include=FALSE}
tab_model(model_FFD1,model_FFD2,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("heat_zone","temp"),
          file="output/tables/Table_models_FFD.html",
          title="Models FFD")
```

### Plots predicted effects

#### Heat zone (model1)

```{r}
plot(ggpredict(model_FFD1,terms=c("heat_zone")),add.data=F)
plot(ggpredict(model_FFD1,terms=c("heat_zone")),add.data=T)
```

There is plasticity in flowering time: individuals planted in warmer areas flower earlier.

#### Temp at planting site (model2)

Non-linear effect?

```{r}
plot(ggpredict(model_FFD2,terms=c("temp[all]")),add.data=T)+
  xlab("Temperature at planting site")
```

There is plasticity in flowering time: individuals planted in warmer areas flower earlier.

#### Interaction heat zone x temp father (model1)

```{r}
plot(ggpredict(model_FFD1,terms=c("temp_father[all]","heat_zone")),add.data=F)+
  scale_color_manual(values = c("#fecc5c", "#fd8d3c", "#e31a1c"))+
  scale_fill_manual(values = c("#fecc5c", "#fd8d3c", "#e31a1c"))+
  xlab("Temperature of the father")
plot(ggpredict(model_FFD1,terms=c("heat_zone","temp_father[minmax]")),add.data=F)+
  scale_color_manual(values = c("#fecc5c", "#e31a1c"))
```

#### Temp father (model2)

```{r}
plot(ggpredict(model_FFD2,terms=c("temp_father")),add.data=F)+
  xlab("Temperature of the father")
```

I was expecting that this effect would go in the opposite direction, and that plants with fathers from colder origins would flower on average earlier. But maybe we should not really interpret this main effect, because the interaction with temperature at planting site is significant?

#### Interaction temp at planting site x temp of the father (model2)

```{r}
plot(ggpredict(model_FFD2,terms=c("temp_father[all]","temp[7.9:49.7 by=5]")),add.data=F,ci=F)+
  scale_color_brewer(palette="YlOrRd")+
  xlab("Temperature of the father")
plot(ggpredict(model_FFD2,terms=c("temp[all]","temp_father[minmax]")),add.data=F,ci=F)+
  scale_color_manual(values = c("#fecc5c", "#e31a1c"))+
  xlab("Temperature at planting site")
```

Plants with fathers from colder origins flower on average earlier when planted on warmer areas,
and on average later when planted on colder areas. Plasticity varies between temperatures of origin of the father: the differences in FFD between planting sites are larger for plants with fathers from colder origin than for plants with fathers from warmer origin.

## Mean temp parents

```{r}
data_transplants<-data_transplants%>%
  mutate(mean_temp_parents=(temp_mother+temp_father)/2)
```

```{r}
model_FFD1_mean<-lmer(FFD_corr~mean_temp_parents*heat_zone+
                   (1|mother)+(1|father)+(1|crossing)+(1|plot),data_transplants)
model_FFD2_mean<-lmer(FFD_corr~mean_temp_parents*temp+
                   (1|mother)+(1|father)+(1|crossing)+(1|plot),data_transplants)
```

```{r}
tab_model(model_FFD1_mean,model_FFD2_mean,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("heat_zone","temp"),
          title="Models FFD mean temp parents")
```

Save models as HTML table

```{r include=FALSE}
tab_model(model_FFD1_mean,model_FFD2_mean,
          transform=NULL,show.ci=F,show.se=T,show.stat=T,digits=3,
          dv.labels=c("heat_zone","temp"),
          file="output/tables/Table_models_FFD_meantemp.html",
          title="Models FFD mean temp parents")
```

### Plots predicted effects

#### Heat zone (model1)

```{r}
plot(ggpredict(model_FFD1_mean,terms=c("heat_zone")),add.data=F)
plot(ggpredict(model_FFD1_mean,terms=c("heat_zone")),add.data=T)
```

#### Temp at planting site (model2)

Non-linear effect?

```{r}
plot(ggpredict(model_FFD2_mean,terms=c("temp[all]")),add.data=T)+
  xlab("Temperature at planting site")
```

#### Interaction heat zone x mean temp parents (model1)

```{r}
plot(ggpredict(model_FFD1_mean,terms=c("mean_temp_parents[all]","heat_zone")),add.data=F)+
  scale_color_manual(values = c("#fecc5c", "#fd8d3c", "#e31a1c"))+
  scale_fill_manual(values = c("#fecc5c", "#fd8d3c", "#e31a1c"))+
  xlab("Mean temperature of parents")
plot(ggpredict(model_FFD1_mean,terms=c("heat_zone","mean_temp_parents[minmax]")),add.data=F)+
  scale_color_manual(values = c("#fecc5c", "#e31a1c"))
```

#### Mean temp parents (model2)

```{r}
plot(ggpredict(model_FFD2_mean,terms=c("mean_temp_parents")),add.data=F)+
  xlab("Mean temperature of the parents")
```

#### Interaction temp at planting site x mean temp of the parents (model2)

```{r}
plot(ggpredict(model_FFD2_mean,terms=c("mean_temp_parents","temp[7.9:49.7 by=5]")),
     add.data=F,ci=F)+
  scale_color_brewer(palette="YlOrRd")+
  xlab("Mean temperature of the parents")
plot(ggpredict(model_FFD2_mean,terms=c("temp","mean_temp_parents[minmax]")),add.data=F,ci=F)+
  scale_color_manual(values = c("#fecc5c", "#e31a1c"))+
  xlab("Temperature at planting site")
```

# Models similar to greenhouse paper

## Prediction 1

### Models

```{r}
model1_FFD<-lmer(FFD_corr~1+(1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot),
                 data_transplants)
model1_LFD<-lmer(LFD_corr~1+(1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=2.459e-07
                      # to avoid singular fit
                   (1|plot),
                 data_transplants)
model1_MeanFD<-lmer(MeanFD~1+(1|father)+(1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|plot),
                 data_transplants)
summary(model1_FFD)
summary(model1_LFD)
summary(model1_MeanFD)
```

### Proportions of variance

```{r}
Variance_FFD<-as.data.frame(VarCorr(model1_FFD))[,c(1,4)]
Variance_LFD<-as.data.frame(VarCorr(model1_LFD))[,c(1,4)]
Variance_MeanFD<-as.data.frame(VarCorr(model1_MeanFD))[,c(1,4)]
# Intra-class correlations
PropVar_FFD <- Variance_FFD%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="FFD")
PropVar_LFD <- Variance_LFD%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="LFD")
PropVar_MeanFD <- Variance_MeanFD%>%mutate(propvar=vcov/sum(vcov))%>%
  mutate(variable="MeanFD")
Props_var<-rbind(PropVar_FFD,PropVar_LFD,PropVar_MeanFD)
```

```{r}
ggplot(Props_var,aes(x=variable,y=propvar,fill=grp))+
  geom_col()
```

### Heritability and maternal effects

```{r}
# h^2 (paternal effects)
her_FFD<-as.numeric(4*subset(PropVar_FFD,grp=="father")[3])
her_LFD<-as.numeric(4*subset(PropVar_LFD,grp=="father")[3])
her_MeanFD<-as.numeric(4*subset(PropVar_MeanFD,grp=="father")[3])
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her<-data.frame(value=rbind(her_FFD,her_LFD,her_MeanFD))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","LFD","MeanFD"),
         effect="Heritability")
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

# But what about the interaction?

mat_FFD<-((Variance_FFD%>%filter(grp=="mother"))$vcov-
            (Variance_FFD%>%filter(grp=="father"))$vcov)/
  as.numeric(Variance_FFD%>%summarise(sum(vcov)))
mat_LFD<-((Variance_LFD%>%filter(grp=="mother"))$vcov-
            (Variance_LFD%>%filter(grp=="father"))$vcov)/
  as.numeric(Variance_LFD%>%summarise(sum(vcov)))
mat_MeanFD<-((Variance_MeanFD%>%filter(grp=="mother"))$vcov-
            (Variance_MeanFD%>%filter(grp=="father"))$vcov)/
  as.numeric(Variance_MeanFD%>%summarise(sum(vcov)))

mat<-data.frame(value=rbind(mat_FFD))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD"),
         effect="Maternal effects")
mat<-data.frame(value=rbind(mat_FFD,mat_LFD,mat_MeanFD))%>%
  rownames_to_column()%>%
  mutate(variable=c("FFD","LFD","MeanFD"),
         effect="Maternal effects")
```

```{r}
her_mat<-rbind(her,mat)
her_mat
```

```{r}
ggplot(her_mat,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")
```

### Table SM: LRTs for variance component

### Table SM: LRTs for variance components

(These results could be used to add asterisks to the previous plot).

```{r}
ranova(model1_FFD)
ranova(model1_LFD)
ranova(model1_MeanFD)
```

## Prediction 2

### Models (linear)

```{r}
model2_FFD<-lmer(FFD_corr~temp+(1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot)+(1|temp:father)+(1|temp:mother),
                 data_transplants)
model2_LFD<-lmer(LFD_corr~temp+
                   #(1|father)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot)+
                   #(1|temp:father)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|temp:mother),
                 data_transplants)
model2_MeanFD<-lmer(MeanFD~temp+
                      #(1|father)+ # Not included because variance=6.776e-05
                      # to avoid singular fit
                      (1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|plot)+(1|temp:father)+(1|temp:mother),
                 data_transplants)
summary(model2_FFD)
summary(model2_LFD)
summary(model2_MeanFD)
```

### Models (quadratic)

```{r}
data_transplants$temp_square<-(data_transplants$temp)^2
model2_FFD_q<-lmer(FFD_corr~temp+temp_square+(1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                     (1|plot)+
                     (1|temp:father)+(1|temp:mother)+
                     (1|temp_square:father)+(1|temp_square:mother),
                   data_transplants)
# Some warnings, not sure what to do
model2_LFD_q<-lmer(LFD_corr~temp+temp_square+
                     #(1|father)+ # Not included because variance=0
                     # to avoid singular fit
                     (1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                     (1|plot)+
                     (1|temp:father)+(1|temp:mother)+
                     (1|temp_square:father)+(1|temp_square:mother),
                   data_transplants)
# Some warnings, not sure what to do
model2_MeanFD_q<-lmer(MeanFD~temp+temp_square+
                        #(1|father)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                     (1|plot)+
                     (1|temp:father)+(1|temp:mother)+
                     (1|temp_square:father)+(1|temp_square:mother),
                   data_transplants)
summary(model2_FFD_q)
summary(model2_LFD_q)
summary(model2_MeanFD_q)
```

### Models (quadratic only in fixed effects)

```{r}
model2_FFD_qf<-lmer(FFD_corr~temp+I(temp^2)+
                      # same as temp+temp_square but better for ggpredict
                      (1|father)+(1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|plot)+(1|temp:father)+(1|temp:mother),
                    data_transplants)
model2_LFD_qf<-lmer(LFD_corr~temp+I(temp^2)+
                      # same as temp+temp_square but better for ggpredict
                      # (1|father)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|plot)+(1|temp:father)+(1|temp:mother),
                    data_transplants)
model2_MeanFD_qf<-lmer(MeanFD~temp+I(temp^2)+
                         # same as temp+temp_square but better for ggpredict
                         # (1|father)+ # Not included because variance=0
                         # to avoid singular fit
                         (1|mother)+
                         #(1|father:mother)+ # Not included because variance=0
                         # to avoid singular fit
                         (1|plot)+(1|temp:father)+(1|temp:mother),
                       data_transplants)
summary(model2_FFD_qf)
summary(model2_LFD_qf)
summary(model2_MeanFD_qf)
```

### Compare linear and quadratic

```{r}
r2_nakagawa(model2_FFD)
r2_nakagawa(model2_FFD_q)
r2_nakagawa(model2_FFD_qf)
r2_nakagawa(model2_LFD)
r2_nakagawa(model2_LFD_q)
r2_nakagawa(model2_LFD_qf)
r2_nakagawa(model2_MeanFD)
r2_nakagawa(model2_MeanFD_q)
r2_nakagawa(model2_MeanFD_qf)
# Conditional R2 (fixed + random effects) only slightly lower in q and qf models
# Marginal R2 (fixed effects) much higher in q and qf models
```

```{r}
AIC(model2_FFD,model2_FFD_q,model2_FFD_qf)
AIC(model2_LFD,model2_LFD_q,model2_LFD_qf)
AIC(model2_MeanFD,model2_MeanFD_q,model2_MeanFD_qf)
# qf models always best --> USE
```

### Proportions of variance

Using models with quadratic effect of temp only as fixed effect

```{r}
# extract variance components
Variance_FFD_temp <- as.data.frame(VarCorr(model2_FFD_qf))[,c(1,4)]  
Variance_LFD_temp <- as.data.frame(VarCorr(model2_LFD_qf))[,c(1,4)]  
Variance_MeanFD_temp <- as.data.frame(VarCorr(model2_MeanFD_qf))[,c(1,4)]  
# Intra-class correlation
PropVar_FFD_temp <- Variance_FFD_temp%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="FFD")
PropVar_LFD_temp <- Variance_LFD_temp%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="LFD")
PropVar_MeanFD_temp <- Variance_MeanFD_temp%>%
  mutate(propvar=vcov/sum(vcov))%>%mutate(variable="MeanFD")
# Proportional variance
Props_var_temp<-rbind(PropVar_FFD_temp,PropVar_LFD_temp,PropVar_MeanFD_temp)
```

```{r}
ggplot(Props_var_temp,aes(x=variable,y=propvar,fill=grp))+
  geom_col()
```

### Heritability and maternal effects

Using models with quadratic effect of temp only as fixed effect

```{r}
# h^2 (paternal effects)
her_FFD_temp<-4*subset(PropVar_FFD_temp,grp=="temp:father")[3]
her_LFD_temp<-0
her_MeanFD_temp<-4*subset(PropVar_MeanFD_temp,grp=="temp:father")[3]
# Because the additive genetic variance, VA,
# is expected to be four times the among pollen‐donor variance 
# (Falconer & Mackay, 1996; Lynch & Walsh, 1998)
her_temp<-data.frame(value=rbind(her_FFD_temp,her_LFD_temp,her_MeanFD_temp))%>%
  mutate(variable=c("FFD","LFD","MeanFD"),
         effect="Heritability")%>%
  rename(value=propvar)
```

```{r}
# maternal effects

# Maternal - paternal
# ----------------------
# summed effects

# Because the pollen‐recipient variance component contains a combination of
# genetic and environmental effects, we subtracted the additive genetic 
# (pollen donor) component from the pollen-recipient variance component
# before dividing the resulting estimate by VP to estimate 
# m2 (m2 = (Vpollen recipient − Vpollen donor)/VP).

mat_FFD_temp<-(subset(Variance_FFD_temp,grp=="temp:mother")[2]-
                  subset(Variance_FFD_temp,grp=="temp:father")[2])/
  Variance_FFD_temp%>%summarise(sum(vcov))
mat_LFD_temp<-(subset(Variance_LFD_temp,grp=="temp:mother")[2]-
                  0)/
  Variance_LFD_temp%>%summarise(sum(vcov))
mat_MeanFD_temp<-(subset(Variance_MeanFD_temp,grp=="temp:mother")[2]-
                  subset(Variance_MeanFD_temp,grp=="temp:father")[2])/
  Variance_MeanFD_temp%>%summarise(sum(vcov))
mat_temp<-data.frame(value=rbind(mat_FFD_temp,mat_LFD_temp,mat_MeanFD_temp))%>%
  mutate(variable=c("FFD","LFD","MeanFD"),
         effect="Maternal effects")%>%
  rename(value=vcov)
```

```{r}
her_mat_temp<-rbind(her_temp,mat_temp)
her_mat_temp
```

```{r}
ggplot(her_mat_temp,aes(x=variable,y=value,fill=effect))+
  geom_bar(stat="identity",position="dodge")
```

### Table SM: LRTs for variance components

Using models with quadratic effect of temp only as fixed effect

```{r}
ranova(model2_FFD_qf)
ranova(model2_LFD_qf)
ranova(model2_MeanFD_qf)
```

## Figure 1 (Predictions 1-2)

## Table 1 (Predictions 1-2)

## Figure 2 (Prediction 2)

### Linear

Main effect of temperature significant for FFD, LFD and MeanFD.

```{r}
plot(ggpredict(model2_FFD,terms="temp[all]"),add.data=T)
plot(ggpredict(model2_LFD,terms="temp[all]"),add.data=T)
plot(ggpredict(model2_MeanFD,terms="temp[all]"),add.data=T)
```

### Quadratic

```{r}
plot(ggpredict(model2_FFD_qf,terms="temp[all]"),add.data=T)
plot(ggpredict(model2_LFD_qf,terms="temp[all]"),add.data=T)
plot(ggpredict(model2_MeanFD_qf,terms="temp[all]"),add.data=T)
```

## Prediction 3

### Models

```{r}
model3_FFD<-lmer(FFD_corr~temp_father+temp_mother+
                   #(1|father)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot),
                 subset(data_transplants,!is.na(FFD_corr)))
model3_LFD<-lmer(LFD_corr~temp_father+temp_mother+
                   (1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot),
                 subset(data_transplants,!is.na(LFD_corr)))
model3_MeanFD<-lmer(MeanFD~temp_father+temp_mother+
                      (1|father)+(1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      (1|plot),
                    subset(data_transplants,!is.na(MeanFD)))
summary(model3_FFD)
summary(model3_LFD)
summary(model3_MeanFD)
```

```{r}
plot(ggpredict(model3_FFD,terms=c("temp_father[all]")))
```

In FFD NS effectof temp_father (but p=0.075), and the sign goes in the opposite direction as expected! (earlier FFD at a given temperature in individuals with fathers from warmer soils). All other effects of orgin temperatures are NS. 

#### Table 2

### Proportions of variance

Proportion of variance explained by temperature of origin of fathers and mothers = proportion of variance explained by fixed effects: marginal R2.

```{r}
PropVar_FFD_temp_origin<-as.numeric(r2_nakagawa(model3_FFD)$R2_marginal)
PropVar_LFD_temp_origin<-as.numeric(r2_nakagawa(model3_LFD)$R2_marginal)
PropVar_MeanFD_temp_origin<-as.numeric(r2_nakagawa(model3_MeanFD)$R2_marginal)
Props_var_temp_origin<-data.frame(
  values=c(PropVar_FFD_temp_origin,PropVar_LFD_temp_origin,
           PropVar_MeanFD_temp_origin),variable=c("FFD","LFD","MeanFD"))
```

```{r}
ggplot(Props_var_temp_origin,aes(x=variable,y=values))+
  geom_bar(stat="identity",position="dodge")
```

##### Temperature mother and father: partR2

```{r eval=FALSE, include=FALSE}
partR2_model3_FFD<-partR2(model3_FFD,
                             partvars=c("temp_father","temp_mother"),
                             R2_type="marginal",nboot=1000)
save(partR2_model3_FFD, file = "output/models/partR2_model3_FFD.rda")
```

```{r eval=FALSE, include=FALSE}
partR2_model3_LFD<-partR2(model3_LFD,
                             partvars=c("temp_father","temp_mother"),
                             R2_type="marginal",nboot=1000)
save(partR2_model3_LFD, file = "output/models/partR2_model3_LFD.rda")
```

```{r eval=FALSE, include=FALSE}
partR2_model3_MeanFD<-partR2(model3_MeanFD,
                             partvars=c("temp_father","temp_mother"),
                             R2_type="marginal",nboot=1000)
save(partR2_model3_MeanFD, file = "output/models/partR2_model3_MeanFD.rda")
```

```{r}
load("output/models/partR2_model3_FFD.rda")
load("output/models/partR2_model3_LFD.rda")
load("output/models/partR2_model3_MeanFD.rda")
```

```{r}
Props_var_temp_mf<-
  rbind(
    partR2_model3_FFD$R2%>%mutate(variable="FFD"),
    partR2_model3_LFD$R2%>%mutate(variable="LFD"),
    partR2_model3_MeanFD$R2%>%mutate(variable="MeanFD")
    )
```

```{r}
ggplot(Props_var_temp_mf%>%
         filter(term!="temp_father+temp_mother"&term!="Full")%>%
         mutate(term=factor(term,levels=c("temp_father","temp_mother"),
                            labels=c("Temperature father",
                                     "Temperature mother"))),
       aes(x=variable,y=estimate,ymin=CI_lower,ymax=CI_upper,
                             fill=term))+
  geom_bar(stat="identity")
```

#### Figure 3

### Models mid-parental values

```{r}
model3_FFD_midP<-lmer(FFD_corr~mean_temp_parents+
                        #(1|father)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|mother)+
                        #(1|father:mother)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|plot),
                      data_transplants)
model3_LFD_midP<-lmer(LFD_corr~mean_temp_parents+
                        (1|father)+(1|mother)+
                        #(1|father:mother)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|plot),
                      data_transplants)
model3_MeanFD_midP<-lmer(MeanFD~mean_temp_parents+
                           (1|father)+(1|mother)+
                           #(1|father:mother)+ # Not included because variance=0
                           # to avoid singular fit
                           (1|plot),
                         data_transplants)
summary(model3_FFD_midP)
summary(model3_LFD_midP)
summary(model3_MeanFD_midP)
```

Mid-parental temperature always NS.

## Prediction 4

### Models

```{r}
model4_FFD<-lmer(FFD_corr~temp*(temp_father+temp_mother)+
                   (1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot),
                 data_transplants)
model4_LFD<-lmer(LFD_corr~temp*(temp_father+temp_mother)+
                   (1|father)+(1|mother)+
                   #(1|father:mother)+ # Not included because variance=0
                   # to avoid singular fit
                   (1|plot),
                 data_transplants)
model4_MeanFD<-lmer(MeanFD~temp*(temp_father+temp_mother)+
                      (1|father)+(1|mother)+
                      #(1|father:mother)+ # Not included because variance=0
                      # to avoid singular fit
                      +(1|plot),
                    data_transplants)
summary(model4_FFD)
summary(model4_LFD)
summary(model4_MeanFD)
```

Interaction temp:temp_father is significant for FFD. There are temperature-related differences among fathers in the response to temperature of the planting site = Genetic differentiation related to origin temperature of the father in the slope of RNs?

```{r}
plot(ggpredict(model4_FFD,terms=c("temp_father")))
```

The sign of temp_father goes in the opposite direction as expected! (earlier FFD at a given temperature in individuals with fathers from warmer soils). But maybe we should not interpret this main effect, as the interaction with temp is significant??

```{r}
plot(ggpredict(model4_FFD,terms=c("temp_father","temp[quart]")))
```

As predicted: differences in phenology between high and low soil temperatures are predicted to be smaller for plants with fathers from warm soils than for plants with fathers from colder soils. 

#### Table 3

#### Figure 4

### Models mid-parental values

```{r}
model4_FFD_midP<-lmer(FFD_corr~temp*mean_temp_parents+
                        (1|father)+(1|mother)+
                        #(1|father:mother)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|plot),
                      data_transplants)
model4_LFD_midP<-lmer(LFD_corr~temp*mean_temp_parents+
                        (1|father)+(1|mother)+
                        #(1|father:mother)+ # Not included because variance=0
                        # to avoid singular fit
                        (1|plot),
                    data_transplants)
model4_MeanFD_midP<-lmer(MeanFD~temp*mean_temp_parents+
                           (1|father)+(1|mother)+
                           #(1|father:mother)+ # Not included because variance=0
                           # to avoid singular fit
                           (1|plot),
                         data_transplants)
summary(model4_FFD_midP)
summary(model4_LFD_midP)
summary(model4_MeanFD_midP)
```

```{r}
plot(ggpredict(model4_FFD_midP,terms=c("mean_temp_parents")))
plot(ggpredict(model4_LFD_midP,terms=c("mean_temp_parents")))
plot(ggpredict(model4_MeanFD_midP,terms=c("mean_temp_parents"))) # Marginally *
```

The sign of mean_temp_parents goes in the opposite direction as expected! (earlier FFD/LFD/meanFD at a given temperature in individuals with parents from warmer soils). But maybe we should not interpret this main effect, as the interaction with temp is significant??

```{r}
plot(ggpredict(model4_FFD_midP,terms=c("mean_temp_parents","temp[quart]")))
plot(ggpredict(model4_LFD_midP,terms=c("mean_temp_parents","temp[quart]")))
plot(ggpredict(model4_MeanFD_midP,terms=c("mean_temp_parents","temp[quart]"))) # Marginally *
```

As predicted: differences in phenology between high and low soil temperatures are predicted to be smaller for plants with parents from warm soils than for plants with parents from colder soils. 

# Predictions 3 and 4 with breeding values?

# Session info

```{r}
sessionInfo()
```
